Move the geometry OBJ loading out of m3d lib into the converter
