+++ Cargo.toml
@@ -42,7 +42,6 @@ byteorder = "1.0"
 cgmath = "0.15"
 gfx = "0.17"
 log = "0.4"
-png = "0.13"
 rand = "0.4"
 rayon = "0.9"
 ron = "0.4"
@@ -57,3 +56,5 @@ gfx_device_gl = "0.15"
 gfx_window_glutin = "0.20"
 glutin = "0.12"
 image = "0.20"
+obj = "0.9"
+png = "0.13"
