+++ bin/convert/model_obj.rs
@@ -1,9 +1,13 @@
-use m3d::{CollisionQuad, DrawTriangle, Geometry, FullModel, Mesh, Model, Slot, Debrie};
+use m3d::{NORMALIZER, NUM_COLOR_IDS,
+    CollisionQuad, ColorId, DrawTriangle, Geometry, Polygon, Vertex,
+    FullModel, Mesh, Model, Slot, Debrie,
+};
 
+use obj::{IndexTuple, Obj, SimplePolygon};
 use ron;
 
 use std::fs::File;
-use std::io::Write;
+use std::io::{Result as IoResult, Write};
 use std::path::PathBuf;
 
 type RefModel = Model<Mesh<String>, Mesh<String>>;
@@ -17,12 +21,12 @@ pub fn export(full: FullModel, model_path: &PathBuf) {
 
     let model = RefModel {
         body: full.body.map(|geom| {
-            geom.save_obj(dir_path.join(BODY_PATH))
+            save_draw_geometry(&geom, dir_path.join(BODY_PATH))
                 .unwrap();
             BODY_PATH.to_string()
         }),
         shape: full.shape.map(|geom| {
-            geom.save_obj(dir_path.join(SHAPE_PATH))
+            save_collision_geometry(&geom, dir_path.join(SHAPE_PATH))
                 .unwrap();
             SHAPE_PATH.to_string()
         }),
@@ -36,7 +40,7 @@ pub fn export(full: FullModel, model_path: &PathBuf) {
                 wheel.map(|mesh| {
                     mesh.map(|geom| {
                         let name = format!("wheel{}.obj", i);
-                        geom.save_obj(dir_path.join(&name)).unwrap();
+                        save_draw_geometry(&geom, dir_path.join(&name)).unwrap();
                         name
                     })
                 })
@@ -48,12 +52,12 @@ pub fn export(full: FullModel, model_path: &PathBuf) {
             .map(|(i, debrie)| Debrie {
                 mesh: debrie.mesh.map(|geom| {
                     let name = format!("debrie{}.obj", i);
-                    geom.save_obj(dir_path.join(&name)).unwrap();
+                    save_draw_geometry(&geom, dir_path.join(&name)).unwrap();
                     name
                 }),
                 shape: debrie.shape.map(|geom| {
                     let name = format!("debrie{}-shape.obj", i);
-                    geom.save_obj(dir_path.join(&name)).unwrap();
+                    save_collision_geometry(&geom, dir_path.join(&name)).unwrap();
                     name
                 }),
             })
@@ -61,7 +65,7 @@ pub fn export(full: FullModel, model_path: &PathBuf) {
         slots: Slot::map_all(full.slots, |mesh, i| {
             mesh.map(|geom| {
                 let name = format!("slot{}.obj", i);
-                geom.save_obj(dir_path.join(&name)).unwrap();
+                save_draw_geometry(&geom, dir_path.join(&name)).unwrap();
                 name
             })
         }),
@@ -77,8 +81,8 @@ pub fn import(model_path: &PathBuf) -> FullModel {
     let model_file = File::open(model_path).unwrap();
     let model = ron::de::from_reader::<_, RefModel>(model_file).unwrap();
 
-    let resolve_geom_draw = |name| -> Geometry<DrawTriangle> { Geometry::load_obj(dir_path.join(name)) };
-    let resolve_geom_coll = |name| -> Geometry<CollisionQuad> { Geometry::load_obj(dir_path.join(name)) };
+    let resolve_geom_draw = |name| -> Geometry<DrawTriangle> { load_geometry(dir_path.join(name)) };
+    let resolve_geom_coll = |name| -> Geometry<CollisionQuad> { load_geometry(dir_path.join(name)) };
     let resolve_mesh = |mesh: Mesh<String>| { mesh.map(&resolve_geom_draw) };
 
     FullModel {
@@ -101,5 +105,170 @@ pub fn import(model_path: &PathBuf) -> FullModel {
         slots: Slot::map_all(model.slots, |mesh, _| {
             resolve_mesh(mesh)
         }),
+    }
+}
+
+
+fn map_color_id(id: u32) -> ColorId {
+    use std::mem;
+    if id < NUM_COLOR_IDS {
+        unsafe { mem::transmute(id) }
+    } else {
+        println!("Unknown ColorId {:?}", id);
+        ColorId::Reserved
+    }
+}
+
+fn flatten_normal(poly: &[IndexTuple], normals: &[[f32; 3]]) -> [i8; 3] {
+    let n = poly.iter().fold([0f32; 3], |u, IndexTuple(_, _, ni)| {
+        let n = match ni {
+            Some(index) => normals[*index],
+            None => [0.0, 0.0, 0.0],
+        };
+        [u[0] + n[0], u[1] + n[1], u[2] + n[2]]
+    });
+    let m2 = n.iter().fold(0f32, |u, v| u + v*v);
+    let scale = if m2 == 0.0 { 0.0 } else { NORMALIZER / m2.sqrt() };
+    [(n[0] * scale) as i8, (n[1] * scale) as i8, (n[2] * scale) as i8]
+}
+
+fn flatten_pos(poly: &[IndexTuple], positions: &[[f32; 3]]) -> [i8; 3] {
+    let m = poly.iter().fold([0f32; 3], |u, IndexTuple(pi, _, _)| {
+        let p = positions[*pi];
+        [u[0] + p[0], u[1] + p[1], u[2] + p[2]]
+    });
+    [ (m[0] * 0.25) as i8, (m[1] * 0.25) as i8, (m[2] * 0.25) as i8 ]
+}
+
+pub fn save_draw_geometry(geom: &Geometry<DrawTriangle>, path: PathBuf) -> IoResult<()> {
+    let mut dest = File::create(&path).unwrap();
+    for p in geom.positions.iter() {
+        writeln!(dest, "v {} {} {}", p[0], p[1], p[2])?;
+    }
+    writeln!(dest, "")?;
+    for n in geom.normals.iter() {
+        writeln!(
+            dest,
+            "vn {} {} {}",
+            n[0] as f32 / NORMALIZER,
+            n[1] as f32 / NORMALIZER,
+            n[2] as f32 / NORMALIZER
+        )?;
+    }
+    writeln!(dest, "")?;
+
+    let mut mask = 0u32;
+    for p in &geom.polygons {
+        mask |= 1 << p.material[0];
+    }
+    for color_id in 0 .. NUM_COLOR_IDS {
+        if mask & (1 << color_id) == 0 {
+            continue
+        }
+        writeln!(dest, "g {:?}", map_color_id(color_id))?;
+        for p in &geom.polygons {
+            if p.material[0] != color_id {
+                continue
+            }
+            write!(dest, "f")?;
+            for v in &p.vertices {
+                write!(dest, " {}//{}", v.pos + 1, v.normal + 1)?;
+            }
+            writeln!(dest, "")?;
+        }
+    }
+
+    Ok(())
+}
+
+pub fn save_collision_geometry(geom: &Geometry<CollisionQuad>, path: PathBuf) -> IoResult<()> {
+    let mut dest = File::create(&path).unwrap();
+    for p in geom.positions.iter() {
+        writeln!(dest, "v {} {} {}", p[0], p[1], p[2])?;
+    }
+    writeln!(dest, "")?;
+
+    // replace the normals with flat normals
+    for p in geom.polygons.iter() {
+        writeln!(
+            dest,
+            "vn {} {} {}",
+            p.flat_normal[0] as f32 / NORMALIZER,
+            p.flat_normal[1] as f32 / NORMALIZER,
+            p.flat_normal[2] as f32 / NORMALIZER
+        )?;
+    }
+    writeln!(dest, "")?;
+
+    for (i, p) in geom.polygons.iter().enumerate() {
+        write!(dest, "f")?;
+        for &pi in &p.vertices {
+            write!(dest, " {}//{}", pi + 1, i + 1)?;
+        }
+        writeln!(dest, "")?;
+    }
+
+    Ok(())
+}
+
+pub fn load_geometry<P: Polygon>(path: PathBuf) -> Geometry<P> {
+    let obj: Obj<SimplePolygon> = Obj::load(&path).unwrap();
+
+    let positions = obj.position
+        .iter()
+        .map(|p| [
+            p[0].min(NORMALIZER).max(-NORMALIZER) as i8,
+            p[1].min(NORMALIZER).max(-NORMALIZER) as i8,
+            p[2].min(NORMALIZER).max(-NORMALIZER) as i8,
+        ])
+        .collect();
+    let normals = obj.normal
+        .iter()
+        .map(|n| [
+            (n[0] * NORMALIZER) as i8,
+            (n[1] * NORMALIZER) as i8,
+            (n[2] * NORMALIZER) as i8,
+        ])
+        .collect();
+
+    let color_names = (0 .. NUM_COLOR_IDS)
+        .map(|id| format!("{:?}", map_color_id(id)))
+        .collect::<Vec<_>>();
+
+    let obj_ref = &obj;
+    let polygons = obj.objects
+        .iter()
+        .flat_map(|object| {
+            object.groups.iter().flat_map(|group| {
+                let mut vertices = Vec::with_capacity(4);
+                let color_id = color_names
+                    .iter()
+                    .position(|c| c == &group.name)
+                    .unwrap_or(0);
+                group.polys
+                    .iter()
+                    .map(move |poly| {
+                        vertices.clear();
+                        for &IndexTuple(pi, _, ni) in poly {
+                            vertices.push(Vertex {
+                                pos: pi as u16,
+                                normal: ni.unwrap_or(0) as u16,
+                            })
+                        }
+                        P::new(
+                            flatten_pos(poly, &obj_ref.position),
+                            flatten_normal(poly, &obj_ref.normal),
+                            [color_id as u32, 0],
+                            &vertices
+                        )
+                    })
+            })
+        })
+        .collect();
+
+    Geometry {
+        positions,
+        normals,
+        polygons,
     }
 }
