+++ lib/m3d/src/geometry.rs
@@ -1,11 +1,3 @@
-use std::io::{self, Write};
-use std::path::PathBuf;
-
-use obj::IndexTuple;
-
-use Polygon;
-
-
 pub const NORMALIZER: f32 = 124.0;
 pub const NUM_COLOR_IDS: u32 = 25;
 
@@ -39,18 +31,6 @@ pub enum ColorId {
     RottenItem = 24,
 }
 
-impl ColorId {
-    pub fn new(id: u32) -> Self {
-        use std::mem;
-        if id < NUM_COLOR_IDS {
-            unsafe { mem::transmute(id) }
-        } else {
-            error!("Unknown ColorId {:?}", id);
-            ColorId::Reserved
-        }
-    }
-}
-
 #[derive(Copy, Clone, Debug)]
 pub struct Vertex {
     pub pos: u16,
@@ -81,170 +61,4 @@ pub struct Geometry<P> {
     pub positions: Vec<[i8; 3]>,
     pub normals: Vec<[i8; 3]>,
     pub polygons: Vec<P>,
-}
-
-fn flatten_normal(poly: &[IndexTuple], normals: &[[f32; 3]]) -> [i8; 3] {
-    let n = poly.iter().fold([0f32; 3], |u, IndexTuple(_, _, ni)| {
-        let n = match ni {
-            Some(index) => normals[*index],
-            None => [0.0, 0.0, 0.0],
-        };
-        [u[0] + n[0], u[1] + n[1], u[2] + n[2]]
-    });
-    let m2 = n.iter().fold(0f32, |u, v| u + v*v);
-    let scale = if m2 == 0.0 { 0.0 } else { NORMALIZER / m2.sqrt() };
-    [(n[0] * scale) as i8, (n[1] * scale) as i8, (n[2] * scale) as i8]
-}
-
-fn flatten_pos(poly: &[IndexTuple], positions: &[[f32; 3]]) -> [i8; 3] {
-    let m = poly.iter().fold([0f32; 3], |u, IndexTuple(pi, _, _)| {
-        let p = positions[*pi];
-        [u[0] + p[0], u[1] + p[1], u[2] + p[2]]
-    });
-    [ (m[0] * 0.25) as i8, (m[1] * 0.25) as i8, (m[2] * 0.25) as i8 ]
-}
-
-impl Geometry<DrawTriangle> {
-    pub fn save_obj(&self, path: PathBuf) -> io::Result<()> {
-        use std::fs::File;
-
-        let mut dest = File::create(&path).unwrap();
-        for p in self.positions.iter() {
-            writeln!(dest, "v {} {} {}", p[0], p[1], p[2])?;
-        }
-        writeln!(dest, "")?;
-        for n in self.normals.iter() {
-            writeln!(
-                dest,
-                "vn {} {} {}",
-                n[0] as f32 / NORMALIZER,
-                n[1] as f32 / NORMALIZER,
-                n[2] as f32 / NORMALIZER
-            )?;
-        }
-        writeln!(dest, "")?;
-
-        let mut mask = 0u32;
-        for p in &self.polygons {
-            mask |= 1 << p.material[0];
-        }
-        for color_id in 0 .. NUM_COLOR_IDS {
-            if mask & (1 << color_id) == 0 {
-                continue
-            }
-            writeln!(dest, "g {:?}", ColorId::new(color_id))?;
-            for p in &self.polygons {
-                if p.material[0] != color_id {
-                    continue
-                }
-                write!(dest, "f")?;
-                for v in &p.vertices {
-                    write!(dest, " {}//{}", v.pos + 1, v.normal + 1)?;
-                }
-                writeln!(dest, "")?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-impl Geometry<CollisionQuad> {
-    pub fn save_obj(&self, path: PathBuf) -> io::Result<()> {
-        use std::fs::File;
-
-        let mut dest = File::create(&path).unwrap();
-        for p in self.positions.iter() {
-            writeln!(dest, "v {} {} {}", p[0], p[1], p[2])?;
-        }
-        writeln!(dest, "")?;
-
-        // replace the normals with flat normals
-        for p in self.polygons.iter() {
-            writeln!(
-                dest,
-                "vn {} {} {}",
-                p.flat_normal[0] as f32 / NORMALIZER,
-                p.flat_normal[1] as f32 / NORMALIZER,
-                p.flat_normal[2] as f32 / NORMALIZER
-            )?;
-        }
-        writeln!(dest, "")?;
-
-        for (i, p) in self.polygons.iter().enumerate() {
-            write!(dest, "f")?;
-            for &pi in &p.vertices {
-                write!(dest, " {}//{}", pi + 1, i + 1)?;
-            }
-            writeln!(dest, "")?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<P: Polygon> Geometry<P> {
-    pub fn load_obj(path: PathBuf) -> Self {
-        use obj::{Obj, SimplePolygon};
-
-        let obj: Obj<SimplePolygon> = Obj::load(&path).unwrap();
-
-        let positions = obj.position
-            .iter()
-            .map(|p| [
-                p[0].min(NORMALIZER).max(-NORMALIZER) as i8,
-                p[1].min(NORMALIZER).max(-NORMALIZER) as i8,
-                p[2].min(NORMALIZER).max(-NORMALIZER) as i8,
-            ])
-            .collect();
-        let normals = obj.normal
-            .iter()
-            .map(|n| [
-                (n[0] * NORMALIZER) as i8,
-                (n[1] * NORMALIZER) as i8,
-                (n[2] * NORMALIZER) as i8,
-            ])
-            .collect();
-
-        let color_names = (0 .. NUM_COLOR_IDS)
-            .map(|id| format!("{:?}", ColorId::new(id)))
-            .collect::<Vec<_>>();
-
-        let obj_ref = &obj;
-        let polygons = obj.objects
-            .iter()
-            .flat_map(|object| {
-                object.groups.iter().flat_map(|group| {
-                    let mut vertices = Vec::with_capacity(4);
-                    let color_id = color_names
-                        .iter()
-                        .position(|c| c == &group.name)
-                        .unwrap_or(0);
-                    group.polys
-                        .iter()
-                        .map(move |poly| {
-                            vertices.clear();
-                            for &IndexTuple(pi, _, ni) in poly {
-                                vertices.push(Vertex {
-                                    pos: pi as u16,
-                                    normal: ni.unwrap_or(0) as u16,
-                                })
-                            }
-                            P::new(
-                                flatten_pos(poly, &obj_ref.position),
-                                flatten_normal(poly, &obj_ref.normal),
-                                [color_id as u32, 0],
-                                &vertices
-                            )
-                        })
-                })
-            })
-            .collect();
-
-        Geometry {
-            positions,
-            normals,
-            polygons,
-        }
-    }
 }
